+++ python/parse_ddl/README.md
@@ -1,6 +1,7 @@
-# Parse SQL
+# Parse DDL
 This project is intended to recursively read through a directory, look into all .sql files and parse out the DDL statements related to schema changes.
 ## Unit Tests
 ```
-python -m unittest test.test_parse_sql -v
+cd ~\Documents\Github\parse_ddl
+python -m unittest tests.test_parse_ddl -v
 ```
