+++ python/parse_ddl/__init__.py
@@ -1,0 +1,1 @@
+"""Initialization of package parse_ddl."""
