+++ python/parse_ddl/sql/PullCustomRaveObjects.sql
@@ -1,6 +1,6 @@
 select * 
 from sys.objects o
-left outer join (select distinct object_name from parse_sql.dbo.parse_sql_ddl) sub on o.name = sub.object_name
+left outer join (select distinct object_name from parse_ddl.dbo.parse_ddl_objects) sub on o.name = sub.object_name
 left outer join (select ViewName from dbo.ClinicalViews) cvs on o.name = cvs.ViewName
 where o.is_ms_shipped = 0
 and sub.object_name is null
