+++ python/parse_ddl/tests/context.py
@@ -1,0 +1,14 @@
+# -*- coding: utf-8 -*-
+"""
+Created on Sat Sep 29 19:56:44 2018
+
+@author: jong
+"""
+
+
+import sys
+import os
+
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
+
+import parse_ddl
