+++ python/parse_ddl/tests/ddl_examples/ddl_function_examples.sql
