+++ python/parse_ddl/tests/ddl_examples/ddl_index_assertions.json
