+++ python/parse_ddl/tests/ddl_examples/ddl_procedure_assertions.json
