+++ python/parse_ddl/tests/ddl_examples/ddl_rename_assertions.json
