+++ python/parse_ddl/tests/ddl_examples/ddl_table_assertions.json
