+++ python/parse_ddl/tests/ddl_examples/ddl_trigger_examples.sql
