+++ python/parse_ddl/tests/ddl_examples/ddl_type_examples.sql
