+++ python/parse_ddl/tests/ddl_examples/ddl_view_examples.sql
