+++ python/parse_ddl/tests/ddl_examples/test_vs.py
@@ -1,0 +1,4 @@
+import json
+import re
+
+print("Hello world")
