+++ python/parse_ddl/tests/file3.sql
