+++ python/parse_ddl/tests/test_logging.py
@@ -1,0 +1,3 @@
+import logging
+logging.basicConfig(filename='log_filename2.txt', level=logging.DEBUG, format='%(asctime)s - %(levelname)s - %(message)s')
+logging.debug('This is a log message.')
