--- python/parse_sql/__init__.py
@@ -1,1 +1,0 @@
-"""Initialization of package parse_sql."""
